requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
pydantic==2.5.1
python-dotenv==1.0.0
web3==6.11.1
cryptography==41.0.5
numpy==1.24.3
random2==1.0.1
jinja2==3.1.5
pandas==2.1.0
pyjwt==2.8.0
watchdog==3.0.0
aiofiles==23.2.1
dask[complete]==2024.1.1
pyarrow==14.0.1
vaex==4.16.0
redis==5.0.1
python-snappy==0.6.1
fastparquet==2024.2.0
distributed==2024.1.1
msgpack==1.0.7
